ignore expected errors on Windows
On Windows, the recv_from operation on the UDP socket may return the
following errors, which are expected and should be ignored:

- 10054 (WSAECONNRESET): Windows can send this error if a previous send
  operation resulted in an ICMP Port Unreachable. And if it's a loopback
  interface, it can know whether there is already another end to
  communicate.
- 10040 (WSAEMSGSIZE): This error was randomly appearing in a test that
  I conducted. Not really sure why it's happening. The frequency
  decreased when I increased the receive buffer size, but it was not
  important to get a network up and running.

Without these changes, it was impossible to get a relatively large
network running without issues. By large I mean a test that might be too
bursting for a single machine to run.

More references:

- http://stackoverflow.com/questions/30749423/is-winsock-error-10054-wsaeconnreset-normal-with-udp-to-from-localhost#comment49588739_30749423
- https://github.com/maidsafe/crust/pull/454